fix: use bigint type for Hyperlane API cursor variable
The Hyperlane GraphQL API expects bigint for id field, not Int.

+++ indexers/indexers/hyperlane_bridge.py
@@ -57,7 +57,7 @@ CHAIN_NAMES = {
 }
 
 HYPERLANE_QUERY = """
-query($cursor: Int!, $limit: Int!) {
+query($cursor: bigint!, $limit: Int!) {
   message_view(
     where: {
       _and: [
